Wrote some tests for MilestoneAwardSelector, moved the hand-crafted 2D array into the test.
This enables, for instance, having a mirrored 2d array internally, makes
things simpler in some ways.
