+++ src/MilestoneAwardSelecter.ts
@@ -34,7 +34,8 @@ import { Tactician } from "./milestones/Tactician";
 import { Terraformer } from "./milestones/Terraformer";
 import { Tycoon } from "./milestones/Tycoon";
 
-const MA_ITEMS = [
+// exported for testing.
+export const MA_ITEMS = [
     ...ORIGINAL_MILESTONES,
     ...ELYSIUM_MILESTONES,
     ...HELLAS_MILESTONES,
@@ -52,8 +53,8 @@ export interface IDrawnMilestonesAndAwards {
 };
 
 export function buildSynergies(): Array<Array<number>> {
-    var array: Array<Array<number>> = new Array(MA_ITEMS.length);
-    for (var idx = 0; idx < MA_ITEMS.length; idx++) {
+    const array: Array<Array<number>> = new Array(MA_ITEMS.length);
+    for (let idx = 0; idx < MA_ITEMS.length; idx++) {
         array[idx] = new Array(MA_ITEMS.length).fill(0)
         array[idx][idx] = 1000;
     }
@@ -61,12 +62,10 @@ export function buildSynergies(): Array<Array<number>> {
     // Higher synergies represent similar milestones or awards. For instance, Terraformer rewards for high TR
     // and the Benefactor award is given to the player with the highets TR. Their synergy weight is 9, very high.
     function bind(first: { new(): IMilestone | IAward }, second: { new(): IMilestone | IAward }, weight: number) {
-        var row = MA_ITEMS.findIndex(ma => new first().name === ma.name);
-        var col = MA_ITEMS.findIndex(ma => new second().name === ma.name);
-        if (row > col) {
-            var tmp = row; row = col; col = tmp;
-        }
+        const row = MA_ITEMS.findIndex(ma => new first().name === ma.name);
+        const col = MA_ITEMS.findIndex(ma => new second().name === ma.name);
         array[row][col] = weight;
+        array[col][row] = weight;
     }
     bind(Terraformer, Benefactor, 9);
     bind(Gardener, Cultivator, 9);
@@ -98,19 +97,19 @@ export function buildSynergies(): Array<Array<number>> {
     bind(Ecologist, Excentric, 4);
     bind(PolarExplorer, Landlord, 4);
     bind(Mayor, Gardener, 3);
-    bind(Gardener, Benefactor, 2);
-    bind(Specialist, Banker, 2);
-    bind(Ecologist, Tycoon, 2);
-    bind(Ecologist, Diversifier, 2);
-    bind(Tycoon, Scientist, 2);
     bind(Tycoon, Excentric, 3);
-    bind(Tycoon, Contractor, 2);
-    bind(Tycoon, Venuphile, 2);
     bind(PolarExplorer, Cultivator, 3);
     bind(Energizer, Thermalist, 3);
     bind(RimSettler, SpaceBaron, 3);
     bind(Celebrity, SpaceBaron, 3);
     bind(Benefactor, Cultivator, 3);
+    bind(Gardener, Benefactor, 2);
+    bind(Specialist, Banker, 2);
+    bind(Ecologist, Tycoon, 2);
+    bind(Ecologist, Diversifier, 2);
+    bind(Tycoon, Scientist, 2);
+    bind(Tycoon, Contractor, 2);
+    bind(Tycoon, Venuphile, 2);
     bind(PolarExplorer, EstateDealer, 2);
     bind(RimSettler, Celebrity, 2);
     bind(Scientist, Magnate, 2);
@@ -147,55 +146,7 @@ export function buildSynergies(): Array<Array<number>> {
     return array;
 }
 
-export const SYNERGIES = [
-    [1000,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,  1,0,0,1,0,0,0,1,1,9,2,0,0,0,0,0],
-    [0,1000,3,0,0,0,0,0,0,0,0,0,4,0,0,0,  6,0,0,0,0,0,0,4,4,0,6,0,0,0,0,0],
-    [0,0,1000,0,0,0,0,1,0,0,0,0,4,0,0,0,  6,0,0,0,0,0,0,4,5,2,9,0,0,0,0,0],
-    [0,0,0,1000,0,0,0,0,4,0,0,0,0,0,0,0,  0,0,0,0,1,0,1,0,0,0,0,5,0,0,9,0],
-    [0,0,0,0,1000,0,0,0,0,0,0,0,0,0,0,0,  0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,1000,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,1000,0,0,0,0,0,0,4,0,0,  0,0,2,1,1,0,1,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,1000,2,0,2,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,1,1,0,4,0,0],
-    [0,0,0,0,0,0,0,0,1000,0,1,1,0,0,1,0,  0,2,0,0,0,0,0,0,0,0,0,5,1,3,2,2],
-    [0,0,0,0,0,0,0,0,0,1000,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,1000,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,1000,0,0,0,0,  0,1,0,0,0,0,0,0,0,0,0,1,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,1000,0,0,0,  4,0,0,0,0,0,0,5,2,0,3,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,1000,0,0,  0,0,0,3,0,0,6,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,1000,0,  0,0,0,0,0,2,0,0,0,0,0,1,3,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1000,  0,0,0,0,0,0,0,0,0,0,0,0,0,5,0,5],
-
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  1000,0,0,0,0,0,0,7,7,0,8,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,1000,0,0,0,0,0,0,0,0,0,2,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,1000,0,0,0,0,0,0,1,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,1000,0,0,0,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,1000,0,7,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,1000,0,0,0,0,0,1,3,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,1000,0,0,0,0,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,1000,5,1,7,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,1000,1,8,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,1000,3,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,1000,0,0,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,1000,2,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,0,1000,0,0,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,0,0,1000,0,2],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,0,0,0,1000,0],
-    [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,  0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1000]
-];
-
-// @ts-ignore unused method except when wanting to dump the array as a list of bindings.
-function _dumpSynergy() {
-    function x(n: number) { return MA_ITEMS[n].constructor.name; }
-    for (var row = 0; row < SYNERGIES.length; row++) {
-        for (var col = 0; col < SYNERGIES[row].length; col++) {
-            var weight = SYNERGIES[row][col];
-            if (weight > 0 && weight < 1000) {
-                console.log(`bind(${x(row)}, ${x(col)}, ${weight});`)
-            }
-        }
-    }
-    console.log("END");    
-}
+const SYNERGIES = buildSynergies();
 
 function shuffleArray(arr: Array<number>) {
     arr = arr.slice()
@@ -214,7 +165,9 @@ function getNumbersRange(start: number, end: number): Array<number> {
     return Array.from(Array(end + 1 - start).keys()).map(n => n + start)
 }
 
-// Selects |count| milestones and |count| awards 
+// Selects |count| milestones and |count| awards from all available awards and milestones (optionally including
+// Venusian.) It does this by way of a weighting algorithm, avoiding high-synergy combinations (such as Gardener
+// and Cultivator, both of which are about placing forest tiles.)
 //
 export function getRandomMilestonesAndAwards(withVenusian: boolean = true, count: number): IDrawnMilestonesAndAwards {
     const maxSynergyAllowed = 1;
@@ -227,7 +180,7 @@ export function getRandomMilestonesAndAwards(withVenusian: boolean = true, count
         // award indexes.
         const rows = shuffleArray(getNumbersRange(0, withVenusian ? 15: 14));
         const cols = shuffleArray(getNumbersRange(16, withVenusian ? 31: 30));
-        candidateIdxs = [...rows.slice(0, count), ...cols.slice(0, count)].sort((a, b) => a - b);
+        candidateIdxs = [...rows.slice(0, count), ...cols.slice(0, count)];
 
         // Search through all pairs of [m1, m2, m3, a1, a2, a3], reading the synergy of
         // each. In this way the maximum synergy is computed between the n*(n-1) pairs.
@@ -243,11 +196,12 @@ export function getRandomMilestonesAndAwards(withVenusian: boolean = true, count
     }
 }
 
-export function computeSynergy(elements: Array<number>) : number {
-    var max = 0;
-    for (let i = 0; i<elements.length - 1; i++) {
-        for (let j = i + 1; j<elements.length; j++) {
-            var synergy = SYNERGIES[elements[i]][elements[j]];
+// Visible for testing.
+export function computeSynergy(indexes: Array<number>) : number {
+    let max = 0;
+    for (let i = 0; i<indexes.length - 1; i++) {
+        for (let j = i + 1; j<indexes.length; j++) {
+            const synergy = SYNERGIES[indexes[i]][indexes[j]];
             max = Math.max(synergy, max);
         }
     }
